requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
fastapi==0.123.0
h11==0.16.0
hiredis==3.3.0
idna==3.11
more-itertools==10.8.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-ulid==1.1.0
redis==5.3.1
redis-om==0.3.5
setuptools==80.9.0
starlette==0.50.0
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
